Fix npm build error
The `npm run build` command failed. The problem is that one of the changes in create-react-app is they added the mini-css-extract-plugin to the production build configuration for webpack (i.e. what is used for yarn build).

he react-app-rewire-create-react-library rewire hasn't been updated for the last year so isn't aware of the change to the webpack 4 configuration, and is removing all plugins from the webpack config.plugins array. This means that the loader for mini-css-extract-plugin has been added for the css but the plugin initialization for mini-css-extract-plugin has been deleted by the rewire, resulting in the error message you are receiving.